+++ app/controllers/application_controller.rb
@@ -4,14 +4,16 @@
 class ApplicationController < ActionController::Base # rubocop:disable Metrics/ClassLength
   include Rails.application.routes.url_helpers
   include ActionView::Helpers::TranslationHelper
+  include Log
+
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
-  protect_from_forgery with: :exception
+  protect_from_forgery with: :exception, prepend: true
 
   before_action :set_locale
   before_action :change_default_caching_policy
 
-  around_action :log_request_result
+  around_action :log_response
 
   # Set the user's preferred locale. An explicit locale set via
   # the URL param `lang` is preeminent, otherwise we look to the
@@ -32,11 +34,22 @@ class ApplicationController < ActionController::Base # rubocop:disable Metrics/C
     expires_in 2.minutes, public: true, must_revalidate: true if Rails.env.production?
   end
 
-  def log_request_result
+  def log_response
     start = Process.clock_gettime(Process::CLOCK_MONOTONIC, :microsecond)
     yield
-    duration = Process.clock_gettime(Process::CLOCK_MONOTONIC, :microsecond) - start
-    detailed_log_result(duration)
+    # Calculate elapsed time and convert to milliseconds
+    duration = (Process.clock_gettime(Process::CLOCK_MONOTONIC, :microsecond) - start) / 1000
+    Log.info(
+      'Processing request',
+      {
+        duration:,
+        method: request.method,
+        params:,
+        path: request.path,
+        request_status: 'processing',
+        status: response.status
+      }
+    )
   end
 
   # Handle specific types of exceptions and render the appropriate error page
@@ -58,14 +71,26 @@ class ApplicationController < ActionController::Base # rubocop:disable Metrics/C
   end
 
   # Render the appropriate error page based on the exception
-  def handle_internal_error(exception)
+  def handle_internal_error(exception) # rubocop:disable Metrics/MethodLength
+    # Render the appropriate error page based on the exception
     if exception.instance_of? ArgumentError
       render_error(400)
     else
-      Rails.logger.warn "No explicit error page for exception #{exception} - #{exception.class}"
-      # Instrument ActiveSupport::Notifications for internal server errors only:
-      sentry_code = instrument_internal_error(exception)
-      render_error(500, sentry_code)
+      cname = exception.class.name
+      logged_fields = {
+        status: Rack::Utils::HTTP_STATUS_CODES[exception]
+      }
+      if Rails.env.development? || Rails.logger.debug?
+        logged_fields[:backtrace] =
+          exception.backtrace.join("\n")
+      end
+      Log.error(
+        "No explicit error page for exception #{exception} - #{cname}",
+        logged_fields
+      )
+      # Instrument ActiveSupport::Notifications for internal errors but only for 500 errors:
+      instrument_application_error(exception)
+      render_error(500)
     end
   end
 
@@ -88,20 +113,20 @@ class ApplicationController < ActionController::Base # rubocop:disable Metrics/C
   def render_error(status, sentry_code = nil)
     reset_response
 
-    status = Rack::Utils::SYMBOL_TO_STATUS_CODE[status] if status.is_a?(Symbol)
+    error_status = Rack::Utils::SYMBOL_TO_STATUS_CODE[status] if status.is_a?(Symbol)
     @view_state ||= LandingState.new(UserLanguageSelection.new(params))
 
     respond_to do |format|
-      format.html { render_html_error_page(status, sentry_code) }
+      format.html { render_html_error_page(error_status, sentry_code) }
       # Anything else returns the status as human readable plain string
-      format.all { render plain: Rack::Utils::HTTP_STATUS_CODES[status].to_s, status: status }
+      format.all { render plain: Rack::Utils::HTTP_STATUS_CODES[status].to_s, status: error_status }
     end
   end
 
   def render_html_error_page(status, sentry_code)
     render 'exceptions/error_page',
-           locals: { status: status, sentry_code: sentry_code },
            layout: true,
+           locals: { status: status, sentry_code: sentry_code },
            status: status
   end
 
@@ -122,56 +147,17 @@ class ApplicationController < ActionController::Base # rubocop:disable Metrics/C
     self.response_body = nil
   end
 
+  def version
+    render json: { version: Version::VERSION }
+  end
+
   private
 
-  # rubocop:disable Metrics/AbcSize, Metrics/MethodLength, Layout/LineLength, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
-  def detailed_log_result(duration)
-    env = request.env
-    query = env['QUERY_STRING'] || URI.parse(env['REQUEST_URI']).query
-    log_fields = {
-      message: response.message || Rack::Utils::HTTP_STATUS_CODES[response.status],
-      path: env['REQUEST_PATH'] || URI.parse(env['REQUEST_URI']).path,
-      request_id: env['X_REQUEST_ID'],
-      request_time: (duration / 1000) || env['REQUEST_TIME'], # in milliseconds
-      method: request.method,
-      status: response.status
-    }
+  def set_sentry_user
+    return unless signed_in? && Rails.env.production?
 
-    log_fields[:path] = "#{log_fields[:path]}?#{query}" if query.present?
-
-    if log_fields[:message] == 'OK' && log_fields[:status] == 200
-      log_fields[:message] = 'Completed request'
-      log_fields[:request_status] = 'completed'
-    end
-
-    if env['HTTP_USER_AGENT'] && Rails.env.production?
-      log_fields[:user_agent] = env['HTTP_USER_AGENT']
-    end
-
-    if (500..599).include?(Rack::Utils::SYMBOL_TO_STATUS_CODE[response.status])
-      log_fields[:message] = env['action_dispatch.exception'].to_s
-      log_fields[:backtrace] = env['action_dispatch.backtrace'].join("\n") unless Rails.env.production?
-    end
-
-    if log_fields[:request_time]
-      log_fields[:message] += format(', time taken: %.0f ms', log_fields[:request_time])
-      seconds, milliseconds = log_fields[:request_time].divmod(1000)
-      log_fields[:request_time] = format('%.0f.%03d', seconds, milliseconds) # rubocop:disable Style/FormatStringToken
-    end
-
-    log_response(response.status, log_fields.sort.to_h)
-  end
-  # rubocop:enable Metrics/AbcSize, Metrics/MethodLength, Layout/LineLength
-
-  # Log the error with the appropriate log level based on the status code
-  def log_response(status, error_log)
-    case status
-    when 500..599
-      Rails.logger.error(JSON.generate(error_log))
-    when 400..499
-      Rails.logger.warn(JSON.generate(error_log))
-    else
-      Rails.logger.info(JSON.generate(error_log))
+    Sentry.configure_scope do |scope|
+      scope.set_user(email: current_user.email)
     end
   end
 
@@ -180,11 +166,11 @@ class ApplicationController < ActionController::Base # rubocop:disable Metrics/C
   # @param [exc] exp the exception that caused the error
   # @return [ActiveSupport::Notifications::Event] provides an object-oriented
   # interface to the event
-  # rubocop:disable Metrics/AbcSize
-  def instrument_internal_error(exc, status = nil)
+  # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
+  def instrument_application_error(exc, status = nil)
     err = {
       message: exc&.message || exc,
-      status: Rack::Utils::SYMBOL_TO_STATUS_CODE[exc]
+      status: exc&.status || Rack::Utils::SYMBOL_TO_STATUS_CODE[exc]
     }
     err[:status] = status if status
     err[:type] = exc.class&.name if exc&.class