+++ app/controllers/compare_controller.rb
@@ -3,7 +3,7 @@
 # Controller for the user action of copmaring statistics between two or more
 # locations
 class CompareController < ApplicationController
-  layout 'webpack_application'
+  layout 'application'
 
   def show
     if params.delete(:print)
@@ -16,7 +16,7 @@ class CompareController < ApplicationController
   private
 
   def setup_view_state
-    LoggingHelper.log_request({ params: params, path: request.path })
+    Log.info('Requesting Compare Controller', { params: params, path: request.path })
     user_compare_selections = UserCompareSelections.new(params)
     query_results = perform_query(user_compare_selections) unless user_compare_selections.search?
 
@@ -27,19 +27,19 @@ class CompareController < ApplicationController
 
   def render_interactive(view_state)
     @view_state = view_state
-    if view_state.respond_to?(:[]) && view_state[:error]
-      render_request_error(@view_state[:user_selections], :internal_server_error) unless Rails.env.development? # rubocop:disable Layout.LineLength
+    if view_state.respond_to?(:[]) && view_state[:error] && Rails.env.production?
+      render_request_error(@view_state[:user_selections], :internal_server_error)
     else
       @view_state
     end
   end
 
   def render_print
-    setup_view_state
+    render_interactive(setup_view_state)
     render 'compare/print', layout: 'print'
   end
 
-  def perform_query(user_compare_selections) # rubocop:disable Metrics/MethodLength, Metrics/AbcSize
+  def perform_query(user_compare_selections)
     query_results = {}
     base_selection = UserSelections.new(
       __safe_params: {
@@ -52,8 +52,7 @@ class CompareController < ApplicationController
       log_fields = { params: base_selection.params, path: request.path }
       msg = 'Received request'
       msg += " for #{location.label}"
-      log_fields[:message] = msg
-      LoggingHelper.log_request(log_fields)
+      Log.info(msg, log_fields)
       query_results[location.label] = query_with(base_selection, location)
     end