+++ app/javascript/components/compare-locations.vue
@@ -1,4 +1,4 @@
-<template lang='html'>
+<template>
   <div class='c-compare__selections'>
     <el-row>
       <el-col :span='24'>
@@ -36,13 +36,13 @@
             {{ location.labels[$locale] || location.labels.en }}
             <button @click='onRemoveLocation(location)' class='c-compare__locations--modify'
               :title="$t('action.remove')" :aria-label="$t('action.remove')">
-              <i class='fa fa-times-circle fa-2x'></i>
+              <i class='fa fa-times-circle'></i>
             </button>
           </li>
           <li v-if='showAddLocationButton' class='c-compare__location'>
             <button class='u-full-width c-compare__locations--modify' @click='onAddLocation'
               :title="$t('action.add_location')" :aria-label="$t('action.add_location')">
-              <i class='fa fa-plus-circle fa-2x'></i>
+              <i class='fa fa-plus-circle'></i>
             </button>
             <compare-additional-location></compare-additional-location>
           </li>
@@ -65,11 +65,11 @@
               {{ $t('js.action.download_data_as') }}
               &nbsp;
             </div>
-            <div>
-              <a :href='downloadUrlCsv' class='c-compare__download-link c-compare__download-csv'>
+            <div class='c-compare__download-links'>
+              <a :href='downloadUrlCsv' class='button c-compare__download-link c-compare__download-csv'>
                 {{ $t('js.compare.csv_format') }} <i class='fa fa-external-link'></i>
               </a>
-              <a :href='downloadUrlJson' class='c-compare__download-link c-compare__download-json'>
+              <a :href='downloadUrlJson' class='button c-compare__download-link c-compare__download-json'>
                 {{ $t('js.compare.json_format') }} <i class='fa fa-external-link'></i>
               </a>
             </div>