+++ test/services/query_command_test.rb
@@ -63,7 +63,10 @@ class QueryCommandTest < ActiveSupport::TestCase
   describe '#query_command?' do
     it 'should report its type correctly' do
       assert qc.query_command?
-      assert_not qc.explain_query_command?
+    end
+
+    it 'should explain its type correctly' do
+      assert qc.explain_query_command?
     end
   end
 end